Hoist the copy-pasted test helpers into the shared common modules
The `codes`/`fires`/`count` (compiler) and `codes(diags)` (LSP e2e) test
helpers were copy-pasted verbatim across many files. Move the canonical,
byte-identical versions into the existing `common` modules (both already
`#![allow(dead_code)]`):

- `tcl-compiler/tests/common/mod.rs` gains `codes`/`fires`/`count`; the
  identical copies in `analyser`, `checks`, `fp_depth`, and
  `compiler_analysis_residual` are replaced by `use common::{…}` (dropping
  the now-unused local imports).
- `tcl-lsp-server/tests/e2e/common/mod.rs` gains `codes(diags)`; the
  identical copies in `diagnostics`, `bigip`, `diagnostic_matrix`, `tcl91`,
  and `edit_tracking_stress` are replaced by `use crate::common::codes;`.

Genuine variants stay put (taint/taint_depth's `warns`-based codes,
alias_scoping's no-dialect form, recovery/vscode_parity's differing e2e
`codes`, and the unrelated `count(ops, op)` bytecode counters). Test-only,
behaviour-identical (531 compiler + 748 e2e tests pass, clippy + fmt clean).
